Major refactor and initial interrupt
