+++ firmware/livex/include/taskComms.h
@@ -1,0 +1,8 @@
+#ifndef TASKCOMMS_H
+#define TASKCOMMS_h
+
+#include <resources.h>
+
+void manageComms();
+
+#endif
