+++ firmware/livex/include/taskPid.h
@@ -1,0 +1,11 @@
+#ifndef TASKPID_H
+#define TASKPID_H
+
+#include <resources.h>
+
+void Core0PIDTask(void * pvParameters);
+void thermalGradient();
+void autoSetPointControl();
+void runPID(String pid);
+
+#endif
